Minor code cleanup in sonar-agent. Use Decimal for thresholds.
Using Double for thresholds that may be checked using equality is error
prone due to the loss of precision and inability to represent certain
whole number values. Decimal is slightly less efficient by does not lose
precision for small values.

+++ sonar-core/Models/MetricHealthCondition.cs
@@ -5,5 +5,5 @@ namespace Cms.BatCave.Sonar.Models;
 
 public record MetricHealthCondition(
   HealthOperator HealthOperator,
-  Double Threshold,
+  Decimal Threshold,
   HealthStatus HealthStatus);
